requirements.txt
astroid==1.6.5
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.0.3
dj-database-url==0.5.0
Django==2.0.7
django-countries==5.3.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
httpie==0.9.9
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.14.5
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
pylint==1.9.2
python-coveralls==2.9.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
wrapt==1.10.11
